_getEditorClass
